appointments/viewsets.py
from rest_framework import viewsets, permissions, filters
from .models import Appointment
from .serializers import AppointmentSerializer


class AppointmentViewSet(viewsets.ModelViewSet):
    queryset = Appointment.objects.all()
    serializer_class = AppointmentSerializer
    permission_classes = [permissions.IsAuthenticated]
    filter_backends = [filters.SearchFilter, filters.OrderingFilter]
    search_fields = ["patient__first_name", "patient__last_name", "reason"]
    ordering_fields = ["date", "start_time", "created_at"]
    ordering = ["-date", "-start_time"]
