requirements.txt
# Django Core
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Image/ PDF Processing
Pillow>=10.1.0
reportlab==4.0.7
xhtml2pdf==0.2.12

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
stripe==7.0.0

# Background Tasks (Celery + Redis)
celery==5.3.4
redis==5.0.1

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
